Add intial support for wall thickness
In order to prepare for doors, and also more interesting looking maps the raycaster can cast slightly into walls to make them optically thinner.

+++ raycaster.h
@@ -9,11 +9,11 @@ class raycaster {
 private:
     int getLookingVerticalOrientation(int lookingDirection);
     int getLookingHorizontalOrientation(int lookingDirection);
-    float* sonarRight(player *player, map *map, float angle);
-    float* sonarLeft(player *player, map *map, float angle);
-    float* sonarUp(player *player, map *map, float angle);
-    float* sonarDown(player *player, map *map, float angle);
-    float calculateRayDistance(int playerX, int playerY, float rayHitX, float rayHitY);
+    int* sonarRight(player *player, map *map, float angle);
+    int* sonarLeft(player *player, map *map, float angle);
+    int* sonarUp(player *player, map *map, float angle);
+    int* sonarDown(player *player, map *map, float angle);
+    float calculateRayDistance(int playerX, int playerY, int rayHitX, int rayHitY);
     float getRadians(float degrees);
     float fixAngleOverflow(float angle);
